requirements.txt
Django==2.2
django-filter==2.1.0
django-filters==0.2.1
djangorestframework==3.9.2
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
